Create Hospital Management System pre-demo_project concept visualization with java
This is a pre-demo based on the preconcept of the main project. It helps in visualization of the preconcept with java showing how the project may work/behave in the background. Since it is a pre-demo based on preconcepts, the final changes which was made it to become the main idea of the main project is not available in this pr- demo. 

Language: Java (for simplicity since it is a pre-demo version)
Data management: text file (for simplicity since it is a pre-demo version)